lib/blobstore: always start extents on new page
This patch is WIP. Idea is to change structure of md to
serializing all of current flags,xattrs to as many pages
as needed.
Then new page for each of the set of extents and
new page for extents table (to be introduced)
that will keep reference to all the extents.

Some patches before this one are missing:
- like keeping set number of extents in single page
- introducing extents table referencing the extents

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: Ifd6597d3e1fb225dc7ae41e161c98506de414d7a