nvme/tcp: Use consistent names for SGL function and variables
Using iovec* as name of functions and variables is good but
iov, iovs, iovcnt have been used widely in SPDK and will be more
concise.

Using common names will be helpful at least to unify related code
into libraries or to follow the best practice in other libraries.

So this patch changes name of functions and related variables.

Besides, we do the following changes.

The name of array points to the first element, and so pointing
array by another pointer is not necessary and can be removed in
nvme_tcp_read_payload_data, nvme_tcp_qpair_process_send_queue, and
spdk_nvmf_tcp_qpair_flush_pdus_internal.

In nvme_tcp_qpair_process_send_queue, the variable mapped_length
is not used and the last parameter of nvme_tcp_build_iovs can
be set to NULL. So remove mapped_length.

Change-Id: I2ad9e6f82136456e9e34bcbb7aca32c5c4d4459a
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>