Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (3 comments)

Hi, Ben.

Name change:  No problem.  Will do.

Removal of Paddr:  Really don't want to do this.  We are testing hardware against "weird addresses."  Some of these are memory that exists but is NOT registered with SPDK.  The way this was implemented is that if the Paddr value is zero, the code calls spdk_vtophys, otherwise it accepts the values as presented.  Is this "negotiable"?  Please?

Combine with create_io_qpair:  Not adverse but didn't want to change the behavior of any existing routines as I don't know who might call them.  Will do if this isn't a concern.

Hope this is how one "replies."  If there's a different mechanism, please advise.

Thanks.
Jim B.

Patch-set: 2
Reviewer: Gerrit User 1015860 <1015860@d5d70762-12d0-45a1-890d-524b12d3f735>