ocf: prevent opening cores during cache load
By default, OCF opens all cores it can find during load procedure.
We don't want that because then we won't
  get reference to a core that was opened this way.
Instead, we want OCF to mark those cores as inactive and then
  add them ourselves.

This patch is a preparation for persistant metadata support.
Changes do not affect current behavior because we dont use load yet.

Change-Id: Ib4c1fe51bdb1c85e66bb61f3598b8bf4895bce57
Signed-off-by: Vitaliy Mysak <vitaliy.mysak@intel.com>