lib/ftl: Add ability to ignore endmeta errors on restore
If a band was padded during recovery, the start md contains
valid data, however end md does not. If the ftl is restarted
again after such a dirty shutdown recovery, and the padded band
hasn't been erased yet, another recovery attempt will find invalid
entry. The config gives the user the ability to specify if any
found inconsistencies in end md should result in an error, or
should be allowed.

Signed-off-by: Mateusz Kozlowski <mateusz.kozlowski@intel.com>
Change-Id: If02f7ca8bc90bb61698fb710fee2274af6af01e4