Update patch set 6
Patch Set 6: Code-Review-1

Hi Changpeng,

100us is ok.  This is only the amount of time until we acknowledge the completion by writing the CQ doorbell.  It does not affect when we process the completions.

From our testing so far, the only time this delay affects command processing is when there are two commands to the same LBA.  But this only exists on some older versions of FW on Cliffdale Refresh.  In that case, the second command will not start until the completion for the first command is acknowledged.  So in this case, the second command may be delayed up to 100us.  But this only affects commands sent to the same LBA which should be very rare.

I'm marking this patch -1 to update the commit message with some of these explanations.

-Jim

Patch-set: 6
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1