blobfs: add more debug logging in sync path
Recent failures related to issue #279 indicate that
a sync was issued for a file but that did not result
in an xattr update in metadata.

I think the only way this can happen is if we see that
the append_pos == length_flushed.  In this case we treat
the sync as a nop - but that's not right, we still need
to write the xattr since the data was likely flushed
implicitly.

But don't fix this just yet - we want to see that this
is exactly the problem causing issue #279, and this ERRLOG
will help prove it.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I764453f74757ef81425d0d9b57548d6c56336da1