nvme/tcp: Correct multiple iovecs case for nvme_tcp_pdu_set_data_buf
nvme_tcp_pdu_set_data_buf has been used to process C2H and H2C for
NVMe/TCP initiator.

In this case, NVMe/TCP cut out the part of the input data buffer
and transfer the part, and repeat these cut and transfer until
the whole data buffer is transferred. NVMe/TCP use two SGLs, and
use one to parse from the offset datao to datao + datal and another
to append from the offset 0 to datal. But current code had not
implemented this case correctly.

This case is very similar with the case covered in DIF library.
This patch takes DIF library as a reference and corrects the
implementation.

The next patch will add UT code.

Change-Id: Ibeed4de182b8b8740566e874e2757280dc21f9e8
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>