Update patch set 1
Patch Set 1: Code-Review+1

Hi Darek,

The commit message is not correct for iSCSI. iSCSI opens LUN and gets a descriptor of the LUN, and use the descriptor to allocate and free the IO channel of the LUN. So the being-reverted change is OK for iSCSI.

The direction we should take is to make vhost/scsi follow iSCSI, and restore this patch then.

Does these match your thought?

Patch is OK to me but the change haven't caused any error in CI and so I want to hear your comment.

-Shuhei

Patch-set: 1
Reviewer: Gerrit User 1012251 <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1