nvmf/tcp: Support single r2t usage
According to the TP 8000 spec in Page 26:
Maximum Number of Outstanding R2T (MAXR2T): Specifies the maximum
number of outstanding R2T PDUs for a command at any point in time
on the connection.

Note that by the spec, the target may max on single r2t
(which is the minimum possible), it doesn't have to use multiple r2ts even
if the initiator supports that. So remove the maxr2t variable in the
tcp qpair structure.

In the original design, we think that maxr2t is the maximal active
active r2t numbers for each connection. So if the initiator sends out
maxr2t=16, it means that all the commands of a qpair can use such number
of R2T pdus. So we need to wait for the available R2Ts for the request
when the maxr2t reaches the maximal vaule. But now, it is different,
each command has its own number of maximal r2t numbers, then do not need to
use the wait method for R2T method anymore. So we remove the state
TCP_REQUEST_STATE_DATA_PENDING_FOR_R2T. Futhermore, we adjust the related
SPDK_TPOINT_ID definition.

Instead, each write command can just send the r2t PDU if needed. And currently
the target will support one active R2T for each command. Thus, we remove the function
spdk_nvmf_tcp_handle_queued_r2t_req.

Reported-by: Or Gerlitz <ogerlitz@mellanox.com>

Signed-off-by: Ziye Yang <ziye.yang@intel.com>
Change-Id: I7547b8facbc39139b4584637ccc51ba8b33ca285