nvmf/tcp: fix the maxr2t usage and support the single R2T PDU first
According to the TP 8000 spec in Page 26:
Maximum Number of Outstanding R2T (MAXR2T): Specifies the maximum
number of outstanding R2T PDUs for a command at any point in time
on the connection.

And this patch eliminates the wrong code to restrict the number
of R2T PDU numbers. And we implement maxr2t=1 in the target,
though the intiator may say that the target can send multiple R2T PDUs
for a write command.

By the way, we may support multiple active r2t pdus for one command later.

Reported-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Ziye Yang <ziye.yang@intel.com>
Change-Id: I7547b8facbc39139b4584637ccc51ba8b33ca285