nvmf/tcp: fix the understanding of maxr2t and support maxr2t=1 first
According to the TP 8000 spec in Page 26:
Maximum Number of Outstanding R2T (MAXR2T): Specifies the maximum
number of outstanding R2T PDUs for a command at any point in time
on the connection.

And this patch eliminates the wrong code, and implement maxr2t=1
behavior, and we will support multiple active r2t for one command later.

Signed-off-by: Ziye Yang <ziye.yang@intel.com>
Change-Id: I7547b8facbc39139b4584637ccc51ba8b33ca285