Update patch set 5
Patch Set 5:

> We have another patch [..] to fix the target side. 

It would be good to put both the initiator and target patches on the same series in gerrit, for clarity and to make sure they are applied together when ready. Otherwise, e.g if only the initiator patch is applied we are broken.

Patch-set: 5
Reviewer: Gerrit User 1016578 <1016578@d5d70762-12d0-45a1-890d-524b12d3f735>