Update patch set 2
Patch Set 2:

Hi Or,

For the target, it does not change the expected behavior, so it will not affect anything.
It removes the wrong usage of maxr2t. Even if we receive maxr2t > 1, our SPDK NVMe/TCP target currently still use 1 R2T for each command, so it will not influence anything. 

According to my test, the performance of big I/O (256KB or 128 KB) and large QD(128). The performance will be improved, since previously there is only 16 R2Ts for each connection, now each command can have 1. so it will not need to wait for R2T.

However, if we want to support multiple R2T, we may still need to wait for R2T for the single command.

Patch-set: 2
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>