nvme: add spdk_nvme_qpair_process_completions_tsc
This will allow transports to make decisions on
coalescing in future patches.  Specifically in mind
is CQ doorbell coalescing for the PCIe transport
which will be done in a future patch.  This patch
just adds the top-level API and the plumbing
to get the tsc value to each of the transport's
completion processing routines.

Note that we allow the user to pass the TSC value
rather than calling spdk_get_ticks() directly to
avoid too many rdtsc instructions which can be
expensive.  With the API implemented here, we can
use the last TSC from the spdk_thread for all of
the qpairs in one iteration.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I194076cde6ac45fc79d4c8f9d768f0e639cef8c5