Update patch set 7
Patch Set 7: Code-Review-1

The bdev layer is designed such that it is supposed to be able to make strong guarantees about the types of I/O it is forwarding down to the bdev modules, and I believe we're already testing that at the bdev layer (if we're not, we should be). The advantage of handling funky I/O sizes in the generic bdev code is that we don't have to handle them in each of the modules, and that lets us remove a lot of repeated code.

Since we're making strong guarantees in the bdev layer that the RAID layer will not see I/O with these sizes, I don't think it makes sense to write a unit test here. The behavior of the RAID module when receiving these I/O sizes can effectively be considered undefined.

Patch-set: 7
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1