Update patch set 5
Patch Set 5:

so the iscsi subsystem uses signed ints and generates several warnings.  With the modules it was as easy as making it clear to the compiler that all were unsigned.  I'm not sure how to fix these or turn off this flag just for iscsi , I tried adding -fno-strict-overflow to the iscsi Makefile but it didn't help.

Any thoughts?

Patch-set: 5
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>