dif: Change the meaning of the data_len parameter of set_md_interleave_iovs
set_md_interleave_iovs() have worked well for iSCSI target, but we have to refine
set_md_interleave_iovs() to use NVMe/TCP target.

There is some difference between iSCSI target and NVMe/TCP target.

- iSCSI target sends R2T PDUs as independent receives multiple Data Out PDUs, and sends
  each Data Out PDU as an independent SCSI write command.
- NVMe/TCP target receives multiple H2C PDUs, and then aggregates multiple H2C PDUs
  into a single NVMe write command.

Accordinto the differnce, to generate and insert DIF into extended LBA payload,
- iSCSI target
 - creates DIF context for each Data Out PDU, and
 - passes the data buffer for each Data Out PDU to set_md_interleave_iovs(),
- NVMe/TCP target
 - creates DIF context once for NVMe write command and then uses the context for
   all H2C PDUs, and
 - passes the whole data buffer for not each H2C PDU but NVMe write command.

As a result,
dif_set_md_interleave_iovs() requires
- iSCSI target
 - the data_offset parameter means the offset in Data Out PDU.
- NVMe/TCP target
 - the data_offset parameter means the offset in not H2C PDU but NVMe write command.

On the other hand,
dif_set_md_interleave_iovs() requires
- iSCSI target
 - the data_len parameter means the end of Data Out PDU
- NVMe/TCP target
 - the data_len parameter means the end of not NVMe write command but H2C PDU.

To satisfy the last requirement, we have to change the meaning of data_len parameter
to be the length of Data Out PDU for iSCSI target and the length of H2C PDU for NVMe/TCP
target.

Hence this patch changes the meaning of data_len parameter to the length of the data,
and changes the implementation accordingly.

Besides, length of Data Out PDU for iSCSI target is ensured to be block multiples, but
length of H2C PDU for NVMe/TCP target is not ensured to be block multiples. Hence this
patch removes the requirement that data_len must be block multiples.

Change-Id: I87722305c299c0356600fc5f4138c5056d6dac9f
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>