Update patch set 8
Patch Set 8: Code-Review-1

> Patch Set 7:
> 
> (1 comment)
> 
> > Patch Set 7: Code-Review-1
> > 
> > The bdev layer is designed such that it is supposed to be able to make strong guarantees about the types of I/O it is forwarding down to the bdev modules, and I believe we're already testing that at the bdev layer (if we're not, we should be). The advantage of handling funky I/O sizes in the generic bdev code is that we don't have to handle them in each of the modules, and that lets us remove a lot of repeated code.
> > 
> > Since we're making strong guarantees in the bdev layer that the RAID layer will not see I/O with these sizes, I don't think it makes sense to write a unit test here. The behavior of the RAID module when receiving these I/O sizes can effectively be considered undefined.
> 
> I would agree for a system test, but this is a UT designed to cover functional code in the RAID module.  Plus, if we assume that the bdev layer is perfect and some bug is introduced at some point that for some reason (no I can't think of one :)) doesn't affect any other module, we'd catch it pretty late.  Personally I think the one extra function in UT is worth it but I'll abandon it if you feel strongly about it. That would leave an assert in the functional code to at least alert of an IO spanning a strip boundary,

I agree that we shouldn't add unit test for conditions that we are guaranteeing won't happen.  If we are worried about a bug getting introduced into the bdev layer, then we should add extra tests for the bdev layer.

Assertions are critical to getting the kind of performance we need in SPDK.  We don't want to check that the caller is always passing in the right thing.  Sure, that's needed on external interfaces like virtio or RPCs, but for internal functions like this, we should rely on these assertions.

Patch-set: 8
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1