test/bdev: measure disk utilization using sample period
NOTE: io sampling functionality is broken right now.

Using bdev sampling the value of disk utilization is calculated.
This works on assumption that device for QOS tests is working
at full capacity for the very first run.
If it was not, then limiting the IOPS might not result
in properly testing it.

Change-Id: I268a1a795218b4ae39655df473826ca161a0f6b0
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>