sock: Add sendmsg API
Added sendmsg to the socket api and posix implementation,
added sendmsg vpp impl that calls internally writev.

Added sock API to get tx zero copy notifications and posix impl.

When the posix socket is created, query the socket provider
if zero copy is supported and if it does, enable that on the
socket. Expose this information to callers through a flag.

Change-Id: I7174c2ed1ac9b1596af162fb0350562718f80e9c
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>