nvmf: Hide bdev's DIF setting if transport's DIF mode is local
When the transport's DIF mode is local, DIF is inserted for
write I/O and stripped for read I/O, and the corresponding
NVMe-oF initiator should not be aware of the DIF setting of the
backend bdev.

Hence this patch hides the DIF setting of the backend bdev
when the DIF mode of the transport is local.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I3c14880c2e94cba7f76b1bca78afb36bfe884e26