Update patch set 8
Patch Set 8:

Hi Paul,

I mean that this test case is OK. However, this case is not running in the CI system due to the test environment restriction. 

And we found a bug with this test case, which should be fixed.

Patch-set: 8
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>